+++ Editor/AssemblyInfo.cs
@@ -1,3 +1,4 @@
 using System.Runtime.CompilerServices;
 
+[assembly: InternalsVisibleTo("com.unity.package-manager-captain.Editor")]
 [assembly: InternalsVisibleTo("com.unity.package-manager-ui.EditorTests")]